{{cookiecutter.app_name}}/requirements.txt
Flask

# Flask Extensions
Flask-Assets
Flask-Login
Flask-Migrate
Flask-Permissions
Flask-Script
Flask-SQLAlchemy
Flask-WTF

# Other
click
pillow
python-dateutil
simplejson
